testing

+++ Jenkinsfile
@@ -2,6 +2,9 @@ pipeline {
 agent any
     stages {
         stage('Build and Running') {
+            when {
+                    branch 'develop'
+                }
             steps {
                 bat 'docker-compose up'
                 // echo 'It works'
@@ -9,6 +12,9 @@ agent any
         }
 
         stage('Running tests') {
+              when {
+                      branch 'develop'
+                  }
                     steps {
 
                   // add steps to run python tests
@@ -20,6 +26,9 @@ agent any
         }
 
         stage('Shutting down') {
+            when {
+                    branch 'develop'
+                }
                     steps {
 
                     bat 'docker-compose down'
